feat(#1079): ICA: lib/extension-state caveman resolvers — circular lib↔caveman seam
